Some progress?

+++ elk/utils/typing.py
@@ -15,7 +15,7 @@ def assert_type(typ: Type[T], obj: Any) -> T:
 
 def float32_to_int16(x: torch.Tensor) -> torch.Tensor:
     """Converts float32 to float16, then reinterprets as int16."""
-    return x.type(torch.float16).view(torch.int16)
+    return x.type(torch.float16).view(torch.float16)
 
 
 def int16_to_float32(x: torch.Tensor) -> torch.Tensor:
